Let the /user EP create users with role=admin
Add the 'role' user claim to tokens. This allows to know if the
authenticated user has administration privileges or not.

If the /user post method is invoked with a token with role=admin, an
additional parameter 'role' is looked up for in the request body. If
it exists and is equal to 'admin', then the new user is created with
role=admin. Otherwise, and by default, users are created with role=user.

Closes #14.