Abort request if token is valid but of a deleted user
If the get method of the /users EP is invoked using a token of a deleted
user, that token is stil valid, and its identity is a user_id that is
no longer in the Users table. But the Users.query.get(user_id) doesn't
throw an exception. It returns None instead. So, the 'No user found'
error was not being shown. That's fixed now.