Fix DB reset for tests
This code was not resetting properly the index of the primary keys.

According to MySql doc:
"You cannot reset the counter to a value less than or equal to the value
that is currently in use. For both InnoDB and MyISAM, if the value is
less than or equal to the maximum value currently in the AUTO_INCREMENT
column, the value is reset to the current maximum AUTO_INCREMENT column
value plus one."

When all data in the table is deleted, its counter has value 0. Then
auto_increment can't be set to a value <=0.